Module aoc_2019::puzzle03

Solution to puzzle03

Contains the code for creating and managing Grids, the structure used for recording the positions through which individual wires have passed, and for using these Grids to solve the two tasks for day 3.

Structs

Grid
Structure for holding the positions of a wire

Enums

Movement
Types of movement

Functions

largestValue
Finds the largest integer in a string.

puzzle1
Runs the solution for the first puzzle.

puzzle2
Runs the solution for the second puzzle.