Struct aoc_2019::puzzle03::Grid

pub struct Grid {
    wires: HashMap<Vec<i64>, i64>,
    current_pos: Vec<i64>,
}

Structure for holding the positions of a wire

Grids are essentially collections of points representing where a wire has passed through; each grid records the positions of a single wire, and can be compared to other grids once populated to find the solution for both of day 3's tasks.

Fields
wires: HashMap<Vec<i64>, i64>current_pos: Vec<i64>
Methods

impl Grid

pub fn grid() -> Grid

Constructor

pub fn get_pos(&self) -> Vec<i64>

Returns this Grid's current_pos.

Returns the current position of this grid's wire as an integer array length 2.

fn move_wire(&mut self, movement: Movement, current_steps: i64) -> i64

Adds another move to this Grid's wire.

Simulates another step in a wire's movement from origin. Takes a Movement enum as input along with a record of the current step count for that wire, adds the new points passed through to this grid's storage, then finally returns the steps taken + 1. This extra step is a bug dealt with in the puzzle solving functions of this module. This could be fixed in future.

Examples

In the below example, a wire movement of upwards 50 is simulated on a brand new Grid.

let mut grid: Grid = Grid::grid();
grid.move_wire(Movement::UP(50));
// Returns "0, 50"
println!("{}, {}", grid.get_pos()[0], grid.get_pos()[1]);

fn execute(&mut self, instructions: &str)

Executes a seriues of instructions on this grid.

Takes a list of instructions in the format U50,D45,R5,etc, parses them with regex, and calls move_wire on this Grid using the appropriate enum.

fn compare(grid1: &Grid, grid2: &Grid) -> i64

Finds the closest interesection of two wires to the origin.

Checks each set of coordinates passed through in grid1 againsts grid2 and if there is an intersection, records whether or not this intersection is closer to the origin than the current best, and finally returns the overall best as an integer.

fn compare_steps(grid1: &Grid, grid2: &Grid) -> i64

Finds the closest interesection of two wires to the origin in terms of distance via wire.

Does exactly the same as compare(), only uses the steps taken by each wire to reach that position (summed together) instead of manhattan distance to the origin to measure 'closeness' to the origin.

pub fn display(&self)

Prints a Grid line by line

Displays a Grid, mostly for debugging purposes.